log: use native syslog facility and severity
Currently when syslog is enabled, AdGuard Home logs using the default
syslog.LOG_NOTICE|syslog.LOG_USER syslog priority, and instead specifies
the log level in a custom-formatted string. Which results in a
duplicated information and less readable syslog entries.

This changes the logging to use the native syslog facility and severity,
not log redundant log level and time, a leaves time keeping to syslog.